requirements.txt
boto3==1.34.47
botocore==1.34.47
jmespath==1.0.1
numpy==1.26.4
pandas==2.2.0
psycopg2-binary==2.9.9
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2024.1
s3transfer==0.10.0
six==1.16.0
tzdata==2024.1
urllib3==2.0.7
